app/requirements.txt
cassandra-driver
pathvalidate
pyarrow
pyspark
